Implement Contains() method
This commit implements the Contains() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.contains?view=netframework-4.8#System_Collections_Generic_List_1_Contains__0_
which is to get the boolean type of information whether the given
element can be found in this Vector.

The Contains() method accepts both basic and custom types, but note that
for custom types operator==() has to be implemented, as it's used to
compare the elements.

Unit tests are added for both basic and custom types.