Implement CopyTo() method with all overloads
The CopyTo() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.copyto?view=netframework-4.8
is made to copy elements from the Vector to destination array given as
an argument.
Each overload adds the possibility to specify starting index of source
or destination data.

Each overload is covered with unit tests checks the basic type only.
(The custom type handling is exactly the same as basic).