Implement FindLast() method
The FindLast() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.findlast?view=netframework-4.8
is responsible for finding the last element in the container, which
matches the criteria specified in the predicate<bool(T)> passed as an
argument.
Unit tests cover fail and success scenarios for both basic and custom
types. Fail scenario should end with the default of template type.