Implement RemoveRange() method
The RemoveRange() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.removerange?view=netframework-4.8
is responsible for removing all elements from the Vector that can be
found in the specified range.
The range is given by starting index and the number of elements to
remove, so the implementation checks whether those constraints fit into
the size of Vector container.
The std::invalid_argument exception is threw when the given range to
remove exceeds the size of container.
Unit tests checks success scenario for both basic and custom type, also
the exception scenario is checked for incorrect arguments.