Implement IndexOf() method with all overloads
The IndexOf() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.indexof?view=netframework-4.8
is responsible for finding the index of very first element that is equal
to the element given as an argument of this method. This implies that
the custom type requires the operator== being implemented in order for
the comparison to be made.
The implementation is based on a general private method that allows each
overload to use the same algorithm.
Unit tests checkes each overload for basic and custom type success
scenarios and fail scenario in general.