Implement FindIndex() method with all overloads
The FindIndex() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.findindex?view=netframework-4.8
is responsible for finding the very first element that meets the
criteria specified by the predicate<bool(T)> given as an argument.
This method is similar to IndexOf(), but instead of operator== this
method uses the predicate as lambda defined by the user.
Unit tests cover all scenarios of success for each of the overload and
one fail caused by the incorrect range given as an argument.