Implement Reverse() method with all overloads
The Reverse() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.reverse?view=netframework-4.8
is responsible for reversing elements in the Vector
that are in the given range, or all elements when no range is specified.
Unit tests cover both scenarios of odd and even number of elements in
the range that will be reversed, and also the error checking is covered
when specified range exceeds the size of Vector.