Implement FindLastIndex() method with all overloads
The FindLastIndex() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.findlastindex?view=netframework-4.8
is responsible for finding the very last index of an element specified
by the criteria given in the predicat<bool(T)> passed as an argument.
The overloads of this method allows to trim the search range by
specifying the starting index and/or ending index of the search.
Unit tests covers all scenarios of success and fail for basic type (the
custom type is only the matter of the predicate implementation and will
not affect the FindLastIndex work correctness).
Also the std::invalid_argument exception condition is checked for each
overload.