Implement InsertRange() method
The InsertRange() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.insertrange?view=netframework-4.8
is responsible for inserting the collection of items (the same type as
the container) in the specified index, while all items are moved to
further indexes.
The implementation provided by this commit adds several overloads
instead of one implementation (like presented in the .NET docs). This is
due to lack of an interface for collections in general in C++1x which
would allow to put any kind of collection to be inserted.
Unit tests covers all scenarios with success and checks whether all
std::invalid_argument exceptions are thrown at certain conditions.