Implement GetRange() method
TheGetRange() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.getrange?view=netframework-4.8
is responsible for creating the shortened list
(NOTE: the new list will store the same type and will be of the same
type as the original container)
starting from 'start' index given and ending with 'end' index given as
an argument.
Unit tests cover the scenario of success for basic type (custom type
does not affect the work of this method, as operator[] is used to get
the element) and the scenario of failure when start and/or end index
given are incorrect.