Implement LastIndexOf() method
The LastIndexOf() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.lastindexof?view=netframework-4.8
is responsible for returning the very last index of the element
specified as an argument of this method.
This implementation is similar to the FindLastIndex() method, but the
only difference is that for this implementation the operator == is used
to check whether the current element matches the one specified as an
argument.
Due to that, the unit tests are basicaly the same as for the predicate<>
version, and checks all the success/fail scenario, plus all the
std::invalid_argument exceptions thrown.