Implement ConvertAll() method
The ConvertAll() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.convertall?view=netframework-4.8
is responsible for converting all of the elements in the container to
the type specified as the template deduction, and put it into new
Cx::ExtendedVector container.
The unit test checks only the success scenario for basic type, as the
difference in method result (success/fail) in conversion is only the
matter of converter<Tout(T)> implementation, which is on the side of
user to do it correctly.