Implement RemoveAt() method
The RemoveAt() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.removeat?view=netframework-4.8
is responsible for removing the element which is under the index
specified as the argument of the method.
In fact, removing single element from std::vector<T> is not difficult:
vector::erase(iterator)
but the original method requires an iterator which has to be either
found, calculated, or moved by the offset. So this implementation of
RemoveAt simplifies the usage of an original method.
Unit tests checks both fail and success scenarios for basic type only.