Rename ExtendedVector to Vector
The ExtendedVector class is the last layer in the Cx::Vector project's
architecture and this class is the one to be used by users by including
the templated header.
The project and the library written is Vector, so the class used further
by other users should be the same so it's convenient and comfortable.
This is why both class name and the file itself has to be renamed so the
Vector name is easy to be found and included, without additional
directories and confusing names.