Remove WrappedContainer layer
The Wrapped/Container class
(which responsibility was to wrap the std::vector<T> class)
is unnecessary layer which makes the including difficult and increase
the dependencies in the library.
The most comfortable usage of Cx::Vector<T> is to include only one
header and just use it, while having additional header only wrapping the
std::vector makes it difficult to use and to deploy easily.
This commit removes the header and moves both wrapped constructors into
the Cx::Vector class implementation.