Fix: RemoveAll does not reduce size of container
During the implementation of example it turned out, that the size of
contianer is not reduced when the RemoveAll() method is called and some
elements are removed. This was causing the behavior to be incorrect.
The reason for this defect is the lack of Cx::Vector::end() given as the
ending index of a range to erase. Due to that, when only the iterator
returned by the remove_if() method was used the size was decreased only
by one.