Implement FindAll() method
The FindAll() method:
https://docs.microsoft.com/pl-pl/dotnet/api/system.collections.generic.list-1.findall?view=netframework-4.8
is responsible for selecting all of the items in the Cx::Vector which
match the criteria given in the predicate<bool(T)>.
If there's no such element, the default empty Cx::Vector<T>() container
is returned.
The unit tests for this method checks both success and failure scenarios
for basic types (the custom type is only the matter of predicate
implementation).