Convert Source/ into shared files project directory
The Source/ directory contains the main source of the SharpVector
implementation: Vector.hpp and Vector.cpp (used only to compile the
header only).
It was based on CMake placed in this directory which was causing the
whole project to contain one configuration with multiple subconfigured
parts of the project.

To make the project structure clean, by seperating each component of the
project, the Source/ directory has been converted into the project
itself with it's own VS configuration.

The Source project is a Shared Items Project
which allows to keep the files simple, but to share them between other
components.