Change Source.vcxitems to Source.vcxproj
The Shared Items Project used for the main source of SharpVector
implementation had the only disadvantage:
It was impossible to build it on it's own.
But building the SharpVector implementation only is crucial for ensuring
the stability of the whole project and the development process.

To let it be built as a standalone project, the Source.vcxitems has been
converted into the static library project: Source.vcxproj