Create Tests directory for CppUnitTestFramework
The CppUnitTestFramework will now be used for the unit testing of the
project (both source and examples).
The Tests/ directory contains the project configuration and the
scaffolded Test.cpp implementation file which preents the most basic
example of unit test implementation.

Test/ directory (previously containing the unit tests) has now been
renamed to Test-old/ to mark it as a legacy, but to leave it so the
implementation is not lost.