Convert App into Examples project
The Examples project is responsible for presenting many features and
functionality of the SharpVector.
The idea is to have many single application which will be able to be ran
from the main() method of the Examples.cpp file.

All this should be possible using many source file of each application
implementation, but having its instance in the main() method guarded by
the preprocessor macro set by the compilation configuration.