Correctly locate the vstest.console.exe tool
It turns out that vstest.console.exe tool is not recognized in the
windows-latest runner on GH Actions servers by default.
To solve this, the additional step has been installed which is to find
the proper vstest.console.exe tool and add it to the path.

Great thanks to Daren May for creating the step:
https://github.com/darenm/Setup-VSTest