Create the .nuspec file for the SharpVector package
The .nuspec file is responsible for managing the package's metadata and
it's content.
More about this file can be found in:
https://docs.microsoft.com/en-us/nuget/reference/nuspec

The nuspec configuration for the project adds three required files:
LICENSE, README and Source/Vector.hpp
All of them are placed in the  SharpVector/ directory after unpacking.

NOTE: Version in the package is currently hardcoded to the 1.0.0, so
it have to be bumped when later version is released.