Remove unused includes from Vector source
The unused includes had to be removed to reduce the number of potential
dependencies, the mess in includes by the user and to slightly reduce
the compilation time.

But the Examples application uses string include, which was previously
provided accidencialy by the Vector.hpp.
It now has to be explicitly included in the source - this is good thing,
as it forces to include what is really used.