Provide CopyTo with std::vector argument overload
The CopyTo does now support std::vector as the in argument array.
It has the same overloads options as the raw array pointer type
overload.
Unit tests has been taken from the pointer-oriented tests and adapted to
the std::vector but in addition they now checks for the correct
insertions.