Implement std::array overloads of CopyTo
The std::array is now supported by the CopyTo method.
The usage is simple, even if the CopyTo needs to be templated method,
there's no argument of size required to be given as a parameter or a
template parameter.

Unit tests for this overload are the same as for raw pointer-oriented
method, as this is the same mechanism because the std::array is fixed
size.